moved sample
